# Manager-feeds/manager-server/swagger.openapi.yaml
openapi: 3.0.3
info:
  title: Feeds Manager Server API
  version: 1.0.0
  description: Complete OpenAPI documentation for all mounted manager-server routes.
servers:
  - url: http://localhost:3001
...
components:
  securitySchemes:
    managerCookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
    ErrorResponse:
      type: object
      required: [success, statusCode, message]
      properties:
        success:
          type: boolean
          example: false
        statusCode:
          type: integer
          example: 403
        message:
          type: string
          example: Access denied for this module
...
    MessageResponse:
      type: object
      required: [success, msg]
      properties:
        success:
          type: boolean
          example: true
        msg:
          type: string
...
    FollowerEntry:
      type: object
      properties:
        username:
          type: string
      additionalProperties: true
    UserSummary:
      type: object
      properties:
        _id:
          type: string
        fullName:
          type: string
        username:
          type: string
        display_name:
          type: string
        email:
          type: string
        phone:
          type: string
        dob:
          type: string
          format: date-time
        visibility:
          type: string
          enum: [Public, Private]
        profilePicture:
          type: string
        bio:
          type: string
        type:
          type: string
          enum: [Kids, Normal, Admin]
        followers:
          type: array
          items:
            $ref: '#/components/schemas/FollowerEntry'
        followings:
          type: array
          items:
            $ref: '#/components/schemas/FollowerEntry'
        blockedUsers:
          type: array
          items:
            type: string
        isPremium:
          type: boolean
        coins:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: true
...
    PostPreview:
      type: object
      properties:
        id:
          type: string
        author:
          type: string
        type:
          type: string
          enum: [Reels, Img]
        url:
          type: string
        content:
          type: string
      additionalProperties: true
    ModerationPost:
      type: object
      properties:
        _id:
          type: string
        id:
          type: string
        type:
          type: string
          enum: [Reels, Img]
        content:
          type: string
        url:
          type: string
        likes:
          type: integer
        dislikes:
          type: integer
        isArchived:
          type: boolean
        warnings:
          type: integer
        createdAt:
          type: string
          format: date-time
      additionalProperties: true
    Report:
      type: object
      properties:
        _id:
          type: string
        report_id:
          type: integer
          enum: [1, 2, 3, 4, 5, 6]
        post_id:
          type: string
        report_number:
          type: integer
        user_reported:
          type: string
        status:
          type: string
          example: Pending
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: true
    ReportWithPreview:
      allOf:
        - $ref: '#/components/schemas/Report'
        - type: object
          properties:
            scopeType:
              type: string
              enum:
                - normal_or_kids_post
                - channel_post
                - normal_chat
                - channel_chat
                - normal_or_kids_account
                - channel_account
                - unknown
            postPreview:
              oneOf:
                - $ref: '#/components/schemas/PostPreview'
                - type: 'null'
...
    UpdateReportStatusRequest:
      type: object
      required: [reportId, status]
      properties:
        reportId:
          type: string
        status:
          type: string
          enum: [Pending, Resolved]
    ReasonRequest:
      type: object
      properties:
        reason:
          type: string
...
    ReportsListResponse:
      type: object
      required: [success, reports]
      properties:
        success:
          type: boolean
          example: true
        reports:
          type: array
          items:
            $ref: '#/components/schemas/ReportWithPreview'
    ReportDetailsResponse:
      type: object
      required: [success, report, post]
      properties:
        success:
          type: boolean
          example: true
        report:
          allOf:
            - $ref: '#/components/schemas/Report'
            - type: object
              properties:
                scopeType:
                  type: string
        post:
          oneOf:
            - $ref: '#/components/schemas/PostPreview'
            - type: 'null'
    ModerationUserResponse:
      type: object
      required: [success, user]
      properties:
        success:
          type: boolean
          example: true
        user:
          allOf:
            - $ref: '#/components/schemas/UserSummary'
            - type: object
              properties:
                followersCount:
                  type: integer
                followingsCount:
                  type: integer
    ModerationUserPostsResponse:
      type: object
      required: [success, posts, count]
      properties:
        success:
          type: boolean
          example: true
        posts:
          type: array
          items:
            $ref: '#/components/schemas/ModerationPost'
        count:
          type: integer
    ModerationUserReportsResponse:
      type: object
      required: [success, reportsAgainstUser]
      properties:
        success:
          type: boolean
          example: true
        reportsAgainstUser:
          type: array
          items:
            $ref: '#/components/schemas/Report'
...
paths:
...
  /report/list:
    get:
      tags: [Reports]
      summary: Get reports queue for posts manager
      security:
        - managerCookieAuth: []
      responses:
        '200':
          description: Reports with post preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportsListResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /report/{id}/details:
    get:
      tags: [Reports]
      summary: Get report details by Mongo report id
      security:
        - managerCookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed report payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDetailsResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /report/updateReportStatus:
    post:
      tags: [Reports]
      summary: Update report status and write manager action log
      security:
        - managerCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportStatusRequest'
      responses:
        '200':
          description: Status updated or already same
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /moderation/user/{username}:
    get:
      tags: [Moderation]
      summary: Get moderation view of one user
      security:
        - managerCookieAuth: []
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details with follower/following counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationUserResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /moderation/user/{username}/posts:
    get:
      tags: [Moderation]
      summary: Get latest posts created by a user
      security:
        - managerCookieAuth: []
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User posts list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationUserPostsResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /moderation/user/{username}/reports:
    get:
      tags: [Moderation]
      summary: Get reports filed against a user
      security:
        - managerCookieAuth: []
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reports against user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationUserReportsResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /moderation/post/{id}:
    delete:
      tags: [Moderation]
      summary: Archive/remove a post by Mongo post id
      security:
        - managerCookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Post removal result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /moderation/user/{username}/warn:
    post:
      tags: [Moderation]
      summary: Warn a user account
      security:
        - managerCookieAuth: []
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Warn action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /moderation/user/{username}/ban:
    post:
      tags: [Moderation]
      summary: Ban a user account
      security:
        - managerCookieAuth: []
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Ban action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...